Support for dynamic slack and rocket channels
